requirements.txt
flask==2.0.1
flask-cors==3.0.10
tweepy==4.12.1
python-dotenv==0.19.0
numpy==1.21.2
pandas==1.3.3
scikit-learn==1.0
nltk==3.6.3